PRL4AirSim/requirements.txt
absl-py==0.13.0
aiohttp==3.8.1
aiosignal==1.2.0
airsim==1.6.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens
async-timeout==4.0.2
atari-py==0.2.9
attrs==21.4.0
autobahn==22.6.1
backcall
beautifulsoup4==4.11.1
bleach==5.0.1
box2d-py==2.3.8
cachetools==4.2.2
catkin_tools
catkin-pkg==0.4.24
certifi
cffi==1.14.5
charset-normalizer==2.0.4
click==8.1.1
cloudpickle==2.1.0
cmake-converter==2.2.0
conda-pack==0.6.0
cryptography==37.0.4
cycler==0.10.0
Cython==0.29.23
data==0.4
debugpy==1.6.2
decorator
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
dm-tree==0.1.7
docker-pycreds==0.4.0
docutils==0.18.1
empy==3.3.4
entrypoints==0.4
executing
fastjsonschema==2.15.3
filelock==3.7.0
fonttools==4.33.3
frozenlist==1.3.0
funcsigs==1.0.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glfw==2.1.0
gmpy2
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.43.0
hyperlink==21.0.0
idna==3.2
imageio==2.19.3
importlib-resources==5.7.1
ipython
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi
Jinja2==3.1.2
joblib==1.0.1
jsonschema==4.5.1
kiwisolver
latex==0.7.0
lockfile==0.12.2
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline
mistune==0.8.4
mkl-fft==1.3.0
mkl-random
mkl-service==2.4.0
mpi4py==3.0.3
mpmath==1.2.1
msgpack==1.0.4
msgpack-rpc-python==0.4
multidict==6.0.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.3
numpy==1.22.4
oauthlib==3.1.1
olefile==0.46
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso
pathtools==0.1.2
pexpect
pickleshare
Pillow==9.1.1
platformdirs==2.5.2
prometheus-client==0.14.1
promise==2.3
prompt-toolkit
protobuf==3.20.1
psutil==5.8.0
ptyprocess
pure-eval
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycparser==2.20
pyglet==1.5.15
Pygments
pyparsing
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
ray==1.12.1
requests==2.26.0
requests-oauthlib==1.3.0
rllib==0.0.1
rosdep==0.21.0
rosdistro==0.8.3
rospkg==1.4.0
rsa==4.7.2
scikit-image==0.19.2
scipy==1.8.1
Send2Trash==1.8.0
sentry-sdk==1.5.8
setproctitle==1.2.2
shortuuid==1.0.8
shutilwhich==1.1.0
simpervisor
sip
six
smmap==5.0.0
snakeviz==2.1.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spyne==2.14.0
stable-baselines3==1.5.0
stack-data
sympy
tabulate==0.8.9
tempdir==0.7.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.5.1
termcolor==1.1.0
terminado==0.15.0
tifffile==2022.5.4
tinycss2==1.1.1
torch
torchaudio
torchvision
tornado
traitlets==5.3.0
txaio==22.2.1
typing-extensions
urllib3==1.26.6
virtualenv==20.14.1
vpython==7.6.3
wandb==0.12.11
wcwidth
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.0.1
widgetsnbextension==3.6.1
yarl==1.7.2
yaspin==2.1.0
zipp==3.8.0
